requirements.txt
yt_dlp
openai-whisper
groq
langchain
langchain-openai
langchain-community
# qdrant-client
langchain-qdrant
